+++ lib/src/core/base/widget/appbar/leading_widget.dart
@@ -21,6 +21,7 @@ import 'package:opticore/opticore.dart';
 /// - [backIcon]: The icon to use for the back button. Default is `FontAwesomeIcons.arrowLeft`.
 /// - [iconColor]: The color of the icon. Default is `null`.
 /// - [iconBackgroundColor]: The background color of the icon. Default is `null`.
+/// - [onBack]: Callback function to execute when the back button is pressed.
 class LeadingWidget extends StatelessWidget {
   /// The route helper used for managing navigation in the app.
   final RouteHelper? route;
@@ -34,17 +35,24 @@ class LeadingWidget extends StatelessWidget {
   /// The background color of the icon.
   final Color? iconBackgroundColor;
 
+  /// Callback function to execute when the back button is pressed.
+  final Function()? onBack;
+
   /// Constructor for the `LeadingWidget`.
   ///
   /// - [route]: The `RouteHelper` that provides navigation functionality,
   ///   allowing the widget to perform actions like `pop()` to go back in the navigation stack.
   /// - [backIcon]: The icon to use for the back button. Default is `FontAwesomeIcons.arrowLeft`.
+  /// - [iconColor]: The color of the icon. Default is `null`.
+  /// - [iconBackgroundColor]: The background color of the icon. Default is `null`.
+  /// - [onBack]: Callback function to execute when the back button is pressed.
   const LeadingWidget({
     super.key,
     required this.route,
     this.backIcon,
     this.iconColor,
     this.iconBackgroundColor,
+    this.onBack,
   });
 
   @override
@@ -61,7 +69,8 @@ class LeadingWidget extends StatelessWidget {
               foregroundColor: iconColor,
               backgroundColor: iconBackgroundColor,
             ),
-            onPressed: () => route?.canPop() ?? false ? route!.pop() : null,
+            onPressed:
+                onBack ?? () => route?.canPop() ?? false ? route!.pop() : null,
           ),
         ],
       ),