+++ lib/src/utils/helper/text_input_helper.dart
@@ -1,92 +1,53 @@
 part of '../util_import.dart';
 
-/// A custom text input formatter that capitalizes the first word of a text input.
-///
-/// This formatter ensures that the first letter of the text input is capitalized,
-/// providing consistent and professional text formatting in forms and other input fields.
-/// It only affects the first letter of the entire input while leaving the rest of the text as-is.
-///
-/// **Main Features:**
-/// - Capitalizes the first letter of the text while maintaining the rest of the text as lowercase.
-/// - Simple to integrate into any text field requiring this formatting behavior.
-///
-/// ### Example Usage:
-/// ```dart
-/// TextField(
-///   inputFormatters: [CapitalizeFirstWordFormatter()],
-/// );
-/// ```
-class CapitalizeFirstWordFormatter extends TextInputFormatter {
-  @override
-  TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    if (newValue.text.isEmpty) return newValue;
+/// A utility class to help manage text input formatting.
+class TextInputHelper {
+  /// Handles formatting for Egyptian phone numbers.
+  ///
+  /// This method listens for changes in the [controller]'s text and, if the country
+  /// phone code is '20' (Egypt), removes the leading zero from the phone number
+  /// if it starts with '0'. It also removes non-printing characters (e.g., LTR marks).
+  ///
+  /// The logic applied:
+  /// - If the phone number starts with a '0' and has more than one character,
+  ///   it will remove the leading '0'.
+  /// - Non-printing characters like Left-To-Right marks (LTR) are also removed.
+  ///
+  /// Example usage:
+  /// ```dart
+  /// TextInputHelper.handleEgyptPhoneInput(
+  ///   controller: _phoneController,
+  ///   countryPhoneCode: selectedCountry?.phoneCode ?? '',
+  /// );
+  /// ```
+  ///
+  /// Parameters:
+  /// - [controller]: The `TextEditingController` that controls the phone number field.
+  /// - [countryPhoneCode]: The phone code of the country (e.g., '20' for Egypt).
+  static void handleEgyptPhoneInput({
+    TextEditingController? controller,
+    String? countryPhoneCode,
+  }) {
+    if (controller == null || countryPhoneCode == null) return;
 
-    // Capitalize only the first letter of the entire input
-    String modifiedText = newValue.text[0].toUpperCase() +
-        newValue.text.substring(
-          1,
+    controller.addListener(() {
+      final String text = controller.text;
+
+      // Remove non-printing characters like LTR mark before performing any operations
+      final cleanedText =
+          text.replaceAll(RegExp(r'[\u200E\u200F\u200B\u200C\u200D]'), '');
+
+      // Proceed with the logic after cleaning the text
+      if (countryPhoneCode == '20' &&
+          cleanedText.startsWith('0') &&
+          cleanedText.length > 1) {
+        // Remove the leading zero if it's present
+        controller.text = cleanedText.substring(1);
+
+        controller.selection = TextSelection.fromPosition(
+          TextPosition(offset: controller.text.length),
         );
-
-    return newValue.copyWith(
-      text: modifiedText,
-      selection: newValue.selection,
-    );
-  }
-}
-
-/// A custom text input formatter that converts Arabic numerals to English numerals.
-///
-/// This formatter ensures that any Arabic numerals (`٠`, `١`, `٢`, etc.)
-/// entered in a text field are automatically converted to their corresponding
-/// English numerals (`0`, `1`, `2`, etc.). This is particularly useful for
-/// applications that require consistent number formatting across different locales
-/// or systems.
-///
-/// **Main Features:**
-/// - Converts Arabic numerals in the input text to English numerals.
-/// - Useful for applications that work with internationalized text inputs and require
-///   consistent numerical representations.
-///
-/// ### Example Usage:
-/// ```dart
-/// TextField(
-///   inputFormatters: [ArabicToEnglishNumberFormatter()],
-/// );
-/// ```
-///
-/// ### Conversion Table:
-/// | AR | EN |
-/// |--------|---------|
-/// |   ٠    |    0    |
-/// |   ١    |    1    |
-/// |   ٢    |    2    |
-/// |   ٣    |    3    |
-/// |   ٤    |    4    |
-/// |   ٥    |    5    |
-/// |   ٦    |    6    |
-/// |   ٧    |    7    |
-/// |   ٨    |    8    |
-/// |   ٩    |    9    |
-class ArabicToEnglishNumberFormatter extends TextInputFormatter {
-  @override
-  TextEditingValue formatEditUpdate(
-      TextEditingValue oldValue, TextEditingValue newValue) {
-    String newText = newValue.text;
-
-    // Replace Arabic numerals with English numerals
-    newText = newText
-        .replaceAll('٠', '0')
-        .replaceAll('١', '1')
-        .replaceAll('٢', '2')
-        .replaceAll('٣', '3')
-        .replaceAll('٤', '4')
-        .replaceAll('٥', '5')
-        .replaceAll('٦', '6')
-        .replaceAll('٧', '7')
-        .replaceAll('٨', '8')
-        .replaceAll('٩', '9');
-
-    return newValue.copyWith(text: newText);
+      }
+    });
   }
 }